Fetches story data from the API and displays
Refactored Story to StoryItem and uses Story as a HOC of sorts to fetch
and decided which state to be in between skeleton and story component.
Added tests around loading data, success, error and dead stories and
how they display. Not caching the fetch calls in the SW as yet but that
should be next.