Adds lazy loading
Moved the hooks into their own folder. Created the StoryList component
to load the latest stories and supply that information to the Story
component. Stories are loaded into a generator so we can keep calling
the next id as we load. The lazy loading is a bit aggressive but the
styling should resolve that or can debounce.